requirements.txt
# Archive-RAG Dependencies
# Python 3.11

# Core RAG Dependencies
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
transformers>=4.30.0

# CLI Framework
typer>=0.9.0

# Logging
structlog>=23.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Topic Modeling
gensim>=4.3.0

# Entity Extraction
spacy>=3.6.0

# Utilities
pydantic>=2.0.0
python-dateutil>=2.8.0

